45 implement a mechanism that will forbid admin users to change things about route assignments and vehicles if the vehicle is on its way (#48)
* * (refactor) end_user_model.py is updated with saved_routes accordingly with Selma Hocam's directive.

* * (refactor) end_user_model.py is updated with saved_routes accordingly with Selma Hocam's directive.

* * (refactor) mongodb_connector.py maxPoolSize is increased from default 100 to 300
* (refactor) deleted cache package, because it'll be risky to hold a lot of data on memory. Not safe.

* * (refacotr) removed is_started field from vehicle_model.py because it makes journal logic a bit more complex and unnecessary.

* (refactor) Updated driver_user_service.py accordingly with the vehicle_model.py

* (refactor) Updated README.md accordingly

* * (feat) Introduced vehicle_state_validator.py to prevent vehicle operations during vehicle is on route. Nice layer of logic. Unfortunately it'll increase my latency but it's worth for it.

* (refactor) Introduced a new method in journal_repository.py that returns boolean if there is an active journey for that vehicle.

* (refactor) Updated admin_vehicle_management_route.py with new feat.

* * (tool) Introduced update_mongodb_fields.py for providing "kind of migrations" structure to add or remove fields from collections. It works, I tested this with Mongo Atlas Cluster to remove is_started from existing vehicles.

* * (refactor) Updated admin_vehicle_management_route.py and admin_vehicle_management_route.py with a new response that returns 409 if the driver is on route on that current process.

* (refactor) Simplified the verification crud method in journal_repository.py which returns bool according to the journal status.

* * (refactor) Added ping_db() from mongodb_connector.py into each repository class in order to check DB Health & Availability.

* * (refactor) Removed duplicate route_uuid data from get_all_vehicles method.

* (refactor) Updated response status code to 409 for vehicle modification attempts while on route

* (refactor) Implemented update_admin_user endpoint and service for admin user updates

* (refactor) Enhanced admin user update functionality with email validation and improved logging

* * (refactor): If an admin tries to assing a vehicle that has already been assigned to another driver, the system will not do the operation and will warn the admin about the vehicle assigned to another user with his name and last name.

+++ src/routes/admin_driver_management_route.py
@@ -1,15 +1,14 @@
 import re
 
-from fastapi import APIRouter, Depends, BackgroundTasks
+from fastapi import APIRouter, Depends
 from fastapi.responses import JSONResponse
-from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
-from fastapi.security.api_key import APIKeyHeader
+from starlette.concurrency import run_in_threadpool
 
 from ..common.request_model.auth_route_models import *
 from ..common.response_model.response_model import ResponseModel
 from ..common.util.logger import get_logger
-from ..common.util.admin_key_validator import validate_admin_api_key
 from ..common.util.jwt_validator import jwt_validator
+from ..common.util.vehicle_state_validator import validate_vehicle_state
 from ..common.request_model.admin_driver_management_models import *
 
 from ..service.admin_driver_management_service import admin_management_service
@@ -70,7 +69,7 @@ def add_driver(
     )
 
 @admin_driver_management_router.delete("/delete_driver", tags=["Admin Driver Management"])
-def delete_driver(
+async def delete_driver(
         driver_data: DeleteDriverUserModel,
         is_jwt_valid: bool = Depends(jwt_validator)
 ) -> JSONResponse:
@@ -78,6 +77,12 @@ def delete_driver(
         return JSONResponse(
             status_code=401,
             content=ResponseModel(success=False, message="Invalid JWT", data={},error="").model_dump())
+    is_vehicle_on_route: bool = await validate_vehicle_state(driver_uid=driver_data.uid)
+    if is_vehicle_on_route:
+        return JSONResponse(
+            status_code=409,
+            content=ResponseModel(success=False, message="Araç yolculuk yaparken değişiklik yapamazsınız",
+                                  data={}, error="").model_dump())
     logger.info(f"Delete driver request for phone number: {driver_data.uid}")
     result: bool = admin_management_service.delete_driver(driver_data.uid)
     return JSONResponse(
@@ -91,7 +96,7 @@ def delete_driver(
 )
 
 @admin_driver_management_router.patch("/update_driver_phone_number", tags=["Admin Driver Management"])
-def update_driver_phone_number(
+async def update_driver_phone_number(
         driver_data: UpdateDriverPhoneNumberModel,
         is_jwt_valid: bool = Depends(jwt_validator)
 ) -> JSONResponse:
@@ -99,6 +104,12 @@ def update_driver_phone_number(
         return JSONResponse(
             status_code=401,
             content=ResponseModel(success=False, message="Invalid JWT", data={},error="").model_dump())
+    is_vehicle_on_route: bool = await validate_vehicle_state(driver_uid=driver_data.uid)
+    if is_vehicle_on_route:
+        return JSONResponse(
+            status_code=409,
+            content=ResponseModel(success=False, message="Araç yolculuk yaparken değişiklik yapamazsınız",
+                                  data={}, error="").model_dump())
     logger.info(f"Update driver phone number request for driver uid: {driver_data.uid}")
     result: dict = admin_management_service.update_driver_phone_number(driver_data.uid, driver_data.new_phone_number)
     return JSONResponse(
@@ -120,6 +131,12 @@ async def assign_vehicle_to_driver(
         return JSONResponse(
             status_code=401,
             content=ResponseModel(success=False, message="Invalid JWT", data={},error="").model_dump())
+    is_vehicle_on_route: bool = await validate_vehicle_state(vehicle_uuid=assignment_data.vehicle_uuid)
+    if is_vehicle_on_route:
+        return JSONResponse(
+            status_code=409,
+            content=ResponseModel(success=False, message="Araç yolculuk yaparken değişiklik yapamazsınız",
+                                  data={}, error="").model_dump())
     logger.info(f"Assign vehicle request for driver UID: {assignment_data.driver_uid} and vehicle UUID: {assignment_data.vehicle_uuid}")
     result: dict = await admin_management_service.assign_vehicle_to_driver(
         assignment_data.driver_uid,
@@ -144,6 +161,12 @@ async def remove_vehicle_from_driver(
         return JSONResponse(
             status_code=401,
             content=ResponseModel(success=False, message="Invalid JWT", data={},error="").model_dump())
+    is_vehicle_on_route: bool = await validate_vehicle_state(driver_uid=driver_uid)
+    if is_vehicle_on_route:
+        return JSONResponse(
+            status_code=409,
+            content=ResponseModel(success=False, message="Araç yolculuk yaparken değişiklik yapamazsınız",
+                                  data={}, error="").model_dump())
     logger.info(f"Remove vehicle from driver request for driver UID: {driver_uid}")
     result: dict = await admin_management_service.remove_vehicle_from_driver(driver_uid)
     return JSONResponse(
